Renames and refactors cb queries #127
