+++ react-backend/database/queries/cb/pwd_token_expire.js
@@ -3,7 +3,6 @@ const dbConnection = require('../../dbConnection');
 const checkTokenquery =
   'SELECT EXISTS(SELECT 1 FROM cbusiness WHERE token = $1 AND tokenexpire >= $2)';
 const checkToken = (token, cb) => {
-  console.log('This is checkExpire query');
   const expire = Date.now();
   dbConnection.query(checkTokenquery, [token, expire], (err, res) => {
     if (err) {
